Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/webaudio/the-audio-api/the-waveshapernode-interface/curve-tests-expected.txt


FAIL WaveShaperNode - Testing that -1, 0 and +1 map correctly to curve (with 1:1 correlation) assert_approx_equals: Input=0, Curve=[2,-3,4] >>>  expected -3 +/- 0.01 but got 0.5
FAIL WaveShaperNode - Testing interpolation (where inputs don't correlate directly to curve elements) assert_approx_equals: Input=-0.5, Curve=[2,-3,4] >>>  expected -0.5 +/- 0.01 but got -1.75
FAIL WaveShaperNode - Testing out-of-range inputs (should be mapped to the first/last elements of the curve) assert_approx_equals: Input=-1.5, Curve=[2,-3,4] >>>  expected 2 +/- 0.01 but got 5.75
FAIL WaveShaperNode - Testing a 2-element curve (does not have a middle element) assert_approx_equals: Input=0, Curve=[2,-2] >>>  expected 0 +/- 0.01 but got -2
FAIL WaveShaperNode - Testing a 4-element curve (does not have a middle element) assert_approx_equals: Input=0, Curve=[1,2,4,7] >>>  expected 3 +/- 0.01 but got 4
FAIL WaveShaperNode - Testing a huge curve assert_approx_equals: Input=0, Curve=[TooBigToDisplay (60000 elements)] >>>  expected 8466.20572879921 +/- 0.01 but got 8466.34765625
PASS WaveShaperNode - Testing null curve (should return input values) 
